refactor: Use utility function
