+++ src/main/kotlin/app/revanced/patches/idaustria/detection/root/RootDetectionPatch.kt
@@ -1,14 +1,13 @@
 package app.revanced.patches.idaustria.detection.root
 
 import app.revanced.patcher.data.BytecodeContext
-import app.revanced.patcher.extensions.InstructionExtensions.addInstruction
-import app.revanced.patcher.extensions.InstructionExtensions.addInstructions
 import app.revanced.patcher.patch.BytecodePatch
 import app.revanced.patcher.patch.annotation.CompatiblePackage
 import app.revanced.patcher.patch.annotation.Patch
 import app.revanced.patches.idaustria.detection.root.fingerprints.AttestationSupportedCheckFingerprint
 import app.revanced.patches.idaustria.detection.root.fingerprints.BootloaderCheckFingerprint
-import app.revanced.patches.idaustria.detection.root.fingerprints.RootCheckFingeprint
+import app.revanced.patches.idaustria.detection.root.fingerprints.RootCheckFingerprint
+import app.revanced.util.Utils.returnEarly
 
 @Patch(
     name = "Remove root detection",
@@ -17,16 +16,11 @@ import app.revanced.patches.idaustria.detection.root.fingerprints.RootCheckFinge
 )
 @Suppress("unused")
 object RootDetectionPatch : BytecodePatch(
-    setOf(AttestationSupportedCheckFingerprint, BootloaderCheckFingerprint, RootCheckFingeprint)
+    setOf(AttestationSupportedCheckFingerprint, BootloaderCheckFingerprint, RootCheckFingerprint)
 ) {
-    override fun execute(context: BytecodeContext) {
-        AttestationSupportedCheckFingerprint.result!!.mutableMethod.addInstruction(0, "return-void")
-        BootloaderCheckFingerprint.result!!.mutableMethod.addInstructions(
-            0, """
-            const v0, 0x1
-            return v0
-        """
-        )
-        RootCheckFingeprint.result!!.mutableMethod.addInstruction(0, "return-void")
-    }
+    override fun execute(context: BytecodeContext) = listOf(
+        AttestationSupportedCheckFingerprint,
+        BootloaderCheckFingerprint,
+        RootCheckFingerprint
+    ).returnEarly()
 }
