+++ src/main/kotlin/app/revanced/patches/idaustria/detection/root/fingerprints/AttestationSupportedCheckFingerprint.kt
@@ -1,6 +1,6 @@
 package app.revanced.patches.idaustria.detection.root.fingerprints
 
-import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
+import app.revanced.patcher.fingerprint.MethodFingerprint
 import com.android.tools.smali.dexlib2.AccessFlags
 
 object AttestationSupportedCheckFingerprint : MethodFingerprint(
