+++ src/main/kotlin/app/revanced/patches/idaustria/detection/root/fingerprints/RootCheckFingerprint.kt
@@ -3,7 +3,7 @@ package app.revanced.patches.idaustria.detection.root.fingerprints
 import app.revanced.patcher.fingerprint.MethodFingerprint
 import com.android.tools.smali.dexlib2.AccessFlags
 
-object RootCheckFingeprint : MethodFingerprint(
+object RootCheckFingerprint : MethodFingerprint(
     "V",
     accessFlags = AccessFlags.PUBLIC.value,
     customFingerprint = { methodDef, _ ->
